Refactored Uncommon components
