+++ src/resources/batches/BatchForm.tsx
@@ -2,8 +2,8 @@ import { yupResolver } from '@hookform/resolvers/yup'
 import React from 'react'
 import { ReferenceInput, SelectInput, SimpleForm, TextInput } from 'react-admin'
 import * as yup from 'yup'
-import DatePicker from './DatePicker'
-import FlexBox from './FlexBox'
+import DatePicker from '../../components/DatePicker'
+import FlexBox from '../../components/FlexBox'
 
 const schema = yup.object({
   year_of_receipt: yup.string().required(),
